fix: replaced edit options by three new options
